# contracts/api/marketplace-central.openapi.yaml
openapi: 3.1.0
info:
  title: Marketplace Central API
  version: 2026-07-13
paths:
...
  /catalog/products:
    get:
      summary: List paginated catalog product facts
      operationId: listCatalogProductFacts
      parameters:
        - name: ids
          in: query
          required: false
          schema:
            type: string
          description: >-
            Comma-separated internal_product_ids. Answers the facts for exactly
            these products instead of a keyset page — a caller that already knows
            its product set (the products linked to listings, say) cannot reach
            scattered ids through the cursor without reading the whole catalog.
            Ids the active source does not carry are absent from the response,
            and no next_cursor is emitted. cursor/limit are ignored when present.
            Rejected with 400 unless it is 1..100 positive integers.
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque base64 keyset cursor returned by the previous page.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: erp_source
          in: query
          required: false
          schema:
            type: string
            enum: [xlsx, catalogo_cliente]
          description: >-
            Selects the imported dataset the reader serves — xlsx (Sankhya ERP,
            the default when omitted) or catalogo_cliente (client catalog import).
            An unknown value is rejected with 400; omitting it is byte-stable with
            pre-toggle clients.
        - name: include_all
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: >-
            When true, returns the complete catalog for this request, ignoring the
            configured sellable-assortment filters. The configuration is unchanged.
        - name: Cache-Control
          in: header
          required: false
          schema:
            type: string
          description: Set to no-cache to map FreshnessPolicy.MaxAge to zero.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogProductFactPage'
        "400":
          description: >-
            Invalid query parameter; the error field names which one
            (cursor, limit, erp_source, include_all, or ids).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
        "503":
          description: Catalog source unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
        "504":
          description: Interactive deadline exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
        "500":
          description: Internal fault reading the catalog; not a source outage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
  /catalog/products/counts:
    get:
      summary: Count configured sellable and total catalog products
      operationId: getCatalogAssortmentCounts
      tags: [catalog]
      responses:
        "200":
          description: Catalog assortment counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogAssortmentCounts'
        "400":
          description: >-
            Invalid query parameter; the error field names which one
            (erp_source).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
        "503":
          description: Catalog source unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
        "504":
          description: Interactive deadline exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
        "500":
          description: Internal fault reading the catalog; not a source outage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
  /catalog/products/{id}:
    get:
      summary: Get a single catalog product
      operationId: getCatalogProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanonicalCatalogProduct'
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /catalog/products/search:
    get:
      summary: Search paginated catalog product facts
      operationId: searchCatalogProductFacts
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: >-
            Opaque base64 keyset cursor returned by the previous page of matches.
            Search pages the same way the list read does: a response whose
            next_cursor is set has more matches than the page holds, and omitting
            the cursor asks for the first page.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 50
        - name: erp_source
          in: query
          required: false
          schema:
            type: string
            enum: [xlsx, catalogo_cliente]
          description: >-
            Selects the imported dataset the reader serves — xlsx (Sankhya ERP,
            the default when omitted) or catalogo_cliente (client catalog import).
            An unknown value is rejected with 400; omitting it is byte-stable with
            pre-toggle clients.
        - name: include_all
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: >-
            When true, returns the complete catalog for this request, ignoring the
            configured sellable-assortment filters. The configuration is unchanged.
        - name: Cache-Control
          in: header
          required: false
          schema:
            type: string
          description: Set to no-cache to map FreshnessPolicy.MaxAge to zero.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogProductFactPage'
        "400":
          description: >-
            Invalid query parameter; the error field names which one. A blank or
            missing q is invalid_q — q is required on this route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
        "503":
          description: Catalog source unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
        "504":
          description: Interactive deadline exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
        "500":
          description: Internal fault reading the catalog; not a source outage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPageErrorResponse'
  /catalog/products/{id}/enrichment:
    get:
      summary: Get product enrichment data
      operationId: getProductEnrichment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductEnrichment'
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Upsert product enrichment data
      operationId: upsertProductEnrichment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductEnrichment'
      responses:
        "200":
          description: Enrichment upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductEnrichment'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
...
  schemas:
...
    CanonicalCatalogProduct:
      type: object
      required: [internal_product_id, name, ean, manufacturer_reference, seller_sku, brand_name, product_group_name, cost_amount, price_amount, stock_quantity, ncm, quality_flags]
      properties:
        internal_product_id:
          type: integer
          minimum: 1
        name:
          type: string
        ean:
          type: string
          nullable: true
        manufacturer_reference:
          type: string
          nullable: true
        seller_sku:
          type: string
          nullable: true
        brand_name:
          type: string
          nullable: true
        product_group_name:
          type: string
          nullable: true
        ncm:
          type: string
          nullable: true
          pattern: '^[0-9]{8}$'
        quality_flags:
          type: array
          items:
            $ref: '#/components/schemas/QualityFlag'
        cost_amount:
          $ref: '#/components/schemas/CanonicalNumericSourceFact'
        price_amount:
          $ref: '#/components/schemas/CanonicalNumericSourceFact'
        stock_quantity:
          $ref: '#/components/schemas/CanonicalNumericSourceFact'
...
    CatalogProductFactPage:
      type: object
      required: [items, next_cursor, page_size, as_of]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CatalogProductFact'
        next_cursor:
          type: string
          nullable: true
          description: Opaque base64 keyset cursor; null on the last page.
        page_size:
          type: integer
          minimum: 0
        as_of:
          type: string
          format: date-time
    CatalogPageErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum: [invalid_cursor, invalid_limit, invalid_q, invalid_erp_source,
                 invalid_ids, invalid_include_all, source_unavailable, deadline_exceeded,
                 internal_error]
        allowed_range:
          type: string
          description: >-
            Present when the error has a bounded accepted domain and states it —
            invalid_limit (the inclusive range), invalid_erp_source (the accepted
            source values), invalid_ids (the accepted id count and shape).
...
    ProductEnrichment:
      type: object
      properties:
        product_id:
          type: string
        height_cm:
          type: number
          nullable: true
        width_cm:
          type: number
          nullable: true
        length_cm:
          type: number
          nullable: true
        weight_g:
          type: number
          nullable: true
        suggested_price_amount:
          type: number
          nullable: true
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: >
                Module-specific error code following the pattern MODULE_ENTITY_REASON.
                Known codes include: CATALOG_PRODUCT_NOT_FOUND, CATALOG_ENRICHMENT_INVALID,
                CLASSIFICATIONS_ENTITY_NOT_FOUND, CLASSIFICATIONS_DUPLICATE_NAME,
                MARKETPLACES_ACCOUNT_INVALID, MARKETPLACES_POLICY_INVALID,
                MARKETPLACES_METHOD_NOT_ALLOWED, MARKETPLACES_FEESCHEDULE_PARAM_MISSING,
                MARKETPLACES_DEFINITIONS_FETCH_FAILED, MARKETPLACES_INTERNAL_ERROR,
                PRICING_SIMULATION_INVALID, PRICING_SIMULATION_INVALID_IDENTITY,
                PRICING_SIMULATION_INVALID_BASE_PRICE, PRICING_SIMULATION_INVALID_VALUES,
                PRICING_SIMULATION_NOT_FOUND, PRICING_METHOD_NOT_ALLOWED,
                CONNECTORS_BATCH_NOT_FOUND, CONNECTORS_RETRY_MISSING_PRODUCT,
                INTERNAL_ERROR, INVALID_REQUEST, NOT_FOUND, CONFLICT.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    CatalogAssortmentCounts:
      type: object
      description: Counts for the configured sellable assortment and the full catalog.
      required: [sellable_count, total_count]
      properties:
        sellable_count:
          type: integer
          minimum: 0
        total_count:
          type: integer
          minimum: 0